Continue waitForTransform when start_time == now()
The previous fix broke waitForTransform with non-strictly
increasing clocks as in `rosbag play --clock`.
This does not always happen, but does regularly.

Using >=, this introduces another possible deadlock:
When waitForTransform starts at the very beginning of the bag
and the timeout exceeds the length of the bag.
This is a pretty ridiculous situation though.