Replace legacy python code with appropriate calls to tf2_ros
The original implementation is not fully compatible with tf2, e.g. it
does not subscribe to /static_tf and thusly cannot support the new
static transformations introduced by tf2.

Similar to the C++ code, the Python implementation forwards all calls to
the tf2 Python implementation now, ensuring that all implementation
details of tf2 are properly hidden and the tf library will remain
compatible.