setTransformStatic not implemented and confusing
In the `TransformerROS` class implemented in tf, the method for setting tf is implemented such as:

https://github.com/ros/geometry/blob/00a32d024af476bf50822e6df2fe2ec97765b1a9/tf/src/tf/listener.py#L62-L63

But it can be confusing when using static tfs. Since no `setTransformStatic` method exist, we would "naturally" fallback to the `setTransform`, which is wrong since it doesn't call `_buffer.set_transform_static` and result in a bad graph. 

To be correct, when setting static tfs to a `TransformerROS` instance we would use something like this

```python
transformer = TransformerROS()
transformer._buffer.set_transform_static(my_static_tf, "default_authority")
```

Otherwise, lookup might fail.

Hope you find this useful :)